feat: improve timeout precision for read and write operations (#68) (#69)
- Change int timeout to float for all read operations (readMessage, readFrame, readLoop)
- Add optional write timeout to sendFrame and sendMessage
- Update Consumer::read/readOne to accept float timeout
- Update Producer::send/sendBatch to accept optional write timeout
- Update Producer::waitForConfirms to use float timeout with microtime precision
- Use consistent RuntimeException for all timeout errors
- Add tests for float timeout signatures and write timeout passthrough

All timeout parameters now accept float values (e.g. 0.5 for 500ms)
for sub-second precision. Write operations support optional timeout
to limit blocking time when socket is not ready.

+++ src/Client/Consumer.php
@@ -59,7 +59,7 @@ class Consumer
     /**
      * @return Message[]
      */
-    public function read(int $timeout = 5): array
+    public function read(float $timeout = 5.0): array
     {
         if (empty($this->buffer)) {
             $this->connection->readLoop(maxFrames: 1, timeout: $timeout);
@@ -78,7 +78,7 @@ class Consumer
         return $messages;
     }
 
-    public function readOne(int $timeout = 5): ?Message
+    public function readOne(float $timeout = 5.0): ?Message
     {
         if (empty($this->buffer)) {
             $this->connection->readLoop(maxFrames: 1, timeout: $timeout);
